Isolate asyncRepair test
